+++ packages/apps/docs/src/components/Layout/Redocly/styles.ts
@@ -1,6 +1,7 @@
 import { styled, StyledComponent } from '@kadena/react-components';
 
 import { BasePageGrid } from '../components';
+import { BaseBackground } from '../Full/components/Aside/AsideStyles';
 
 export const PageGrid: StyledComponent<typeof BasePageGrid> = styled(
   BasePageGrid,
@@ -27,6 +28,59 @@ export const PageGrid: StyledComponent<typeof BasePageGrid> = styled(
     '@2xl': {
       gridTemplateColumns:
         'minmax(1%, auto) $leftSideWidth minmax(auto, calc($pageWidth - $leftSideWidth)) minmax(1%, auto)',
+    },
+  },
+);
+
+export const CodeBackground: StyledComponent<typeof BaseBackground> = styled(
+  BaseBackground,
+
+  {
+    $$shadowWidth: '$sizes$20',
+    '&::before': {
+      display: 'none',
+
+      content: '',
+      position: 'absolute',
+      pointerEvents: 'none',
+      inset: 0,
+      top: '$10',
+      backgroundColor: '$background',
+      backgroundImage: 'url("/assets/bg-vertical.png")',
+      backgroundRepeat: 'no-repeat',
+      backgroundPositionY: '-100px',
+      backgroundPositionX: '-100px',
+      '@md': {
+        backgroundColor: 'transparent',
+        backgroundPositionX:
+          'calc(100vw  - ($$asideMenuWidthCode + $$shadowWidth))',
+      },
+      '@lg': {
+        backgroundPositionX:
+          'calc(100vw  - ($$asideMenuWidthCode + $$shadowWidth))',
+      },
+      '@xl': {
+        display: 'block',
+      },
+      '@2xl': {
+        backgroundPositionX:
+          'calc($sizes$pageWidth + ((100vw - $sizes$pageWidth) /2 ) - ($$asideMenuWidthCode + $$shadowWidth))',
+      },
+    },
+    '&::after': {
+      backgroundColor: 'transparent',
+      '@md': {
+        left: 'calc(100vw  - ($$asideMenuWidthCode +  $sizes$4))',
+      },
+      '@lg': {
+        left: 'calc(100vw  - ($$asideMenuWidthCode +  $sizes$4  + $4))',
+      },
+      '@xl': {
+        backgroundColor: '$backgroundOverlayColor',
+      },
+      '@2xl': {
+        left: 'calc($sizes$pageWidth + ((100vw - $sizes$pageWidth) /2) - ($$asideMenuWidthCode + $6 ))',
+      },
     },
   },
 );
